requirements.txt
Flask~=3.0.0
pyodbc~=4.0.39
flask-cors==3.0.10
requests~=2.28.2
python-decouple